# openapi.json
{
  "components": {
    "schemas": {
      "Approval": {
        "additionalProperties": false,
        "properties": {
          "approved_at": {
            "format": "date-time",
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "manual": {
            "type": "boolean"
          },
          "matched_rule": {
            "type": "string"
          },
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "state": {
            "enum": [
              "open",
              "merged",
              "closed",
              "unknown"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "title_parts": {
            "$ref": "#/components/schemas/TitleParts"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "number",
          "title",
          "title_parts",
          "author",
          "url",
          "matched_rule",
          "manual",
          "state",
          "approved_at"
        ],
        "type": "object"
      },
...
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ManualApproveOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ManualApproveOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "number"
        ],
        "type": "object"
      },
...
      "QueueItem": {
        "additionalProperties": false,
        "properties": {
          "additions": {
            "format": "int64",
            "type": "integer"
          },
          "author": {
            "type": "string"
          },
          "changed_files": {
            "format": "int64",
            "type": "integer"
          },
          "deletions": {
            "format": "int64",
            "type": "integer"
          },
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "title_parts": {
            "$ref": "#/components/schemas/TitleParts"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "number",
          "title",
          "title_parts",
          "author",
          "url",
          "additions",
          "deletions",
          "changed_files",
          "reasons"
        ],
        "type": "object"
      },
...
      "TitleParts": {
        "additionalProperties": false,
        "properties": {
          "breaking": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "scopes",
          "breaking",
          "description"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "toilmaster3000",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/toilmaster3000/v1/approvals": {
      "get": {
        "operationId": "list-approvals",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Approval"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Approval Feed, newest-first"
      }
    },
    "/api/toilmaster3000/v1/queue": {
      "get": {
        "operationId": "list-queue",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Needs-Human-Review queue (derived live)"
      }
    },
    "/api/toilmaster3000/v1/queue/{number}/approve": {
      "post": {
        "operationId": "approve-queue-item",
        "parameters": [
          {
            "description": "PR number to approve from the queue",
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "description": "PR number to approve from the queue",
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualApproveOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Manual override approve (the only path to approve a breaking change)"
      }
    },
    "/api/toilmaster3000/v1/queue/{number}/diff": {
      "get": {
        "operationId": "get-queue-diff",
        "parameters": [
          {
            "description": "PR number whose diff to fetch from the queue",
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "description": "PR number whose diff to fetch from the queue",
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PRDiffBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Fetch a queued PR's diff on demand (the Diff card)"
      }
    },
    "/api/toilmaster3000/v1/rules": {
      "get": {
        "operationId": "list-rules",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleBody"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List rules in file order"
      },
      "post": {
        "operationId": "create-rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create a rule"
      }
    },
    "/api/toilmaster3000/v1/rules/{id}": {
      "delete": {
        "operationId": "delete-rule",
        "parameters": [
          {
            "description": "Stable rule id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable rule id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Delete a rule"
      },
      "put": {
        "operationId": "update-rule",
        "parameters": [
          {
            "description": "Stable rule id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable rule id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Full-replace a rule (also enable/disable)"
      }
    },
    "/api/toilmaster3000/v1/status": {
      "get": {
        "operationId": "get-status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleStatus"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Cycle status line"
      }
    }
  }
}